chore(pkg/scripts_test): fix linux lint failures
