+++ pkg/scripts_test/check_linux.go
@@ -1,0 +1,11 @@
+package sumologic_scripts_tests
+
+import "github.com/stretchr/testify/require"
+
+func checkTokenEnvFileCreated(c check) {
+	require.FileExists(c.test, tokenEnvFilePath, "env token file has not been created")
+}
+
+func checkTokenEnvFileNotCreated(c check) {
+	require.NoFileExists(c.test, tokenEnvFilePath, "env token file not been created")
+}
