+++ pkg/scripts_test/common_darwin.go
@@ -1,0 +1,11 @@
+package sumologic_scripts_tests
+
+import (
+	"os"
+	"path"
+)
+
+func getPackagePath() string {
+	tmpDir := os.Getenv("TMPDIR")
+	return path.Join(tmpDir, darwinPackageName)
+}
