+++ pkg/scripts_test/consts_darwin.go
@@ -1,0 +1,6 @@
+package sumologic_scripts_tests
+
+const (
+	darwinPackageName         string = "otelcol-sumo.pkg"
+	darwinUninstallScriptPath string = appSupportDirPath + "/uninstall.sh"
+)
